+++ package.json
@@ -100,11 +100,7 @@
     "test:unit": "jest ./app/ ./locales/ ./tests/**/*.test.ts .github/**/*.test.ts ./scripts/**/*.test.ts --testPathIgnorePatterns='.*/tests/smoke-appium/.*\\.spec\\.(ts|tsx|js)$|.*/e2e/.*\\.spec\\.(ts|tsx|js)$|.*\\.integration\\.test\\.(ts|tsx|js|jsx)$|.*/tests/llm-workflow/.*\\.test\\.(ts|tsx|js)$|.*\\.view(\\..*)?\\.test\\.(ts|tsx|js|jsx)$'",
     "test:unit:llm-workflow": "jest -c jest.config.llm-workflow.js",
     "mm:doctor": "tsx tests/llm-workflow/doctor-entry.ts",
-    "performance:swaps:analyze": "tsx tests/performance/swaps/analyze.ts",
-    "performance:swaps:cleanup": "tsx tests/performance/swaps/instrumentation-cli.ts cleanup",
-    "performance:swaps:prepare": "tsx tests/performance/swaps/instrumentation-cli.ts prepare",
-    "performance:swaps:scenario:001": "tsx tests/performance/swaps/scenarios/001-open-swaps-fetch-one-eth-quote.ts",
-    "performance:swaps:status": "tsx tests/performance/swaps/instrumentation-cli.ts status",
+    "performance:swaps": "tsx tests/performance/swaps/cli/index.ts",
     "test:view:base": "jest -c jest.config.view.js --no-watchman --silent",
     "test:view": "yarn test:view:base --coverage=false --testPathPattern='.*\\.view(\\..*)?\\.test\\.(ts|tsx|js|jsx)$'",
     "test:view:one": "yarn test:view:base --verbose --coverage=false --testTimeout=30000",
