+++ tests/performance/swaps/cli/run-scenario.ts
@@ -1,35 +1,36 @@
 /* eslint-disable import-x/no-nodejs-modules */
+import { spawnSync } from 'node:child_process';
 import { mkdirSync, writeFileSync } from 'node:fs';
 import { resolve } from 'node:path';
-import { spawnSync } from 'node:child_process';
+import { SwapsPerformanceArtifact } from '../analysis/artifact';
+import { formatArtifactMarkdown } from '../analysis/markdown-report';
+import { summarizeCapture } from '../analysis/summarize';
 import {
   buildDrainDiagnosticsExpression,
   buildInstallDiagnosticsExpression,
   buildMarkerExpression,
   extractCdpEvaluationValue,
-  extractInteractionText,
-  formatArtifactMarkdown,
-  hasPositiveNumericValue,
   parseRuntimeCapture,
   RuntimeCapture,
+} from '../capture/hermes-collector';
+import { getInstrumentationStatus } from '../capture/source-instrumentation';
+import { resolveScenario } from '../scenarios/registry';
+import {
+  ScenarioContext,
   ScenarioPhase,
-  summarizeCapture,
-  SWAPS_PERFORMANCE_SCENARIO_001,
-  SwapsPerformanceArtifact,
-} from '../diagnostics';
+  ScenarioPreconditionState,
+} from '../scenarios/types';
 import {
   buildMmSessionProbeArgs,
+  extractInteractionText,
   formatMmSessionSetupCommand,
-} from '../runner-config';
+  parseMetroPort,
+} from './runner-support';
 
 const LOG_PREFIX = '[SWAPS_PERF_ANALYSIS]';
-const DEFAULT_METRO_PORT = 8081;
-const QUOTE_TIMEOUT_MS = 30_000;
-const QUOTE_POLL_INTERVAL_MS = 350;
 const COMMAND_TIMEOUT_MS = 60_000;
 // Maximum timeout accepted by the mm CLI.
 const MM_COMMAND_TIMEOUT_MS = 30_000;
-const WALLET_SWAP_BUTTON_TEST_ID = 'homepage-action-buttons-grid-swap';
 const OUTPUT_DIRECTORY = resolve(
   process.cwd(),
   'test-reports/swaps-performance',
@@ -85,10 +86,7 @@ function parseCommandOutput(output: string): unknown {
 
 function runMm(args: string[], allowFailure = false): unknown {
   const result = runYarn(['mm', ...args], allowFailure);
-  if (!result.ok) {
-    return null;
-  }
-  return parseCommandOutput(result.stdout);
+  return result.ok ? parseCommandOutput(result.stdout) : null;
 }
 
 function evaluateRuntime(expression: string): unknown {
@@ -120,89 +118,12 @@ function readRuntimeCapture(): RuntimeCapture | null {
   }
 }
 
-function getVisibleText(testId: string, allowFailure = false): string | null {
-  const output = runMm(
-    ['get-text', '--testid', testId, '--timeout', '5000'],
-    allowFailure,
-  );
-  return extractInteractionText(output, testId);
-}
-
-function getExactScreenText(testId: string): string | null {
-  const output = runMm(['describe-screen']);
-  return extractInteractionText(output, testId);
-}
-
-function waitForTestId(testId: string, timeoutMs: number): void {
-  runMm(['wait-for', '--testid', testId, '--timeout', String(timeoutMs)]);
-}
-
-function clickTestId(testId: string): void {
-  runMm(['click', '--testid', testId, '--timeout', '10000']);
-}
-
-function delay(durationMs: number): Promise<void> {
-  return new Promise((resolveDelay) => {
-    setTimeout(resolveDelay, durationMs);
-  });
-}
-
-async function waitForPositiveQuote(): Promise<string> {
-  const deadline = Date.now() + QUOTE_TIMEOUT_MS;
-  let lastText: string | null = null;
-
-  while (Date.now() < deadline) {
-    lastText = getVisibleText('dest-token-area-input', true);
-    if (hasPositiveNumericValue(lastText)) {
-      return lastText ?? '';
-    }
-    await delay(QUOTE_POLL_INTERVAL_MS);
-  }
-
-  throw new Error(
-    `First quote did not become visible within ${QUOTE_TIMEOUT_MS}ms; last destination text was ${JSON.stringify(
-      lastText,
-    )}`,
-  );
-}
-
-async function measurePhase(
-  name: ScenarioPhase['name'],
-  action: () => Promise<void> | void,
-): Promise<ScenarioPhase> {
-  markRuntime(`${name}:start`);
-  const startedAt = Date.now();
-  await action();
-  const endedAt = Date.now();
-  markRuntime(`${name}:end`);
-
-  return {
-    name,
-    startedAt,
-    endedAt,
-    durationMs: endedAt - startedAt,
-  };
-}
-
 function getCommit(): string {
   const result = spawnSync('git', ['rev-parse', '--short', 'HEAD'], {
     cwd: process.cwd(),
     encoding: 'utf8',
   });
   return result.status === 0 ? result.stdout.trim() : 'unknown';
-}
-
-function parseMetroPort(argv: string[]): number {
-  const flagIndex = argv.indexOf('--metro-port');
-  if (flagIndex === -1) {
-    return DEFAULT_METRO_PORT;
-  }
-
-  const value = Number(argv[flagIndex + 1]);
-  if (!Number.isInteger(value) || value <= 0 || value > 65_535) {
-    throw new Error('--metro-port must be an integer between 1 and 65535');
-  }
-  return value;
 }
 
 function writeArtifact(artifact: SwapsPerformanceArtifact): {
@@ -217,16 +138,53 @@ function writeArtifact(artifact: SwapsPerformanceArtifact): {
   return { jsonPath, markdownPath };
 }
 
-async function runScenario(): Promise<void> {
-  const metroPort = parseMetroPort(process.argv.slice(2));
+function createScenarioContext(): ScenarioContext {
+  return {
+    log,
+    clickTestId: (testId) => {
+      runMm(['click', '--testid', testId, '--timeout', '10000']);
+    },
+    waitForTestId: (testId, timeoutMs) => {
+      runMm(['wait-for', '--testid', testId, '--timeout', String(timeoutMs)]);
+    },
+    getVisibleText: (testId, allowFailure = false) => {
+      const output = runMm(
+        ['get-text', '--testid', testId, '--timeout', '5000'],
+        allowFailure,
+      );
+      return extractInteractionText(output, testId);
+    },
+    getExactScreenText: (testId) =>
+      extractInteractionText(runMm(['describe-screen']), testId),
+    delay: (durationMs) =>
+      new Promise((resolveDelay) => {
+        setTimeout(resolveDelay, durationMs);
+      }),
+    now: Date.now,
+    measurePhase: async (name, action): Promise<ScenarioPhase> => {
+      markRuntime(`${name}:start`);
+      const startedAt = Date.now();
+      await action();
+      const endedAt = Date.now();
+      markRuntime(`${name}:end`);
+      return { name, startedAt, endedAt, durationMs: endedAt - startedAt };
+    },
+  };
+}
+
+export async function runSwapsPerformanceScenario(
+  scenarioReference: string,
+  scenarioArgs: string[],
+): Promise<void> {
+  const scenario = resolveScenario(scenarioReference);
+  const metroPort = parseMetroPort(scenarioArgs);
   const createdAt = new Date();
-  const runId = `${SWAPS_PERFORMANCE_SCENARIO_001.id.toLowerCase()}-${
-    SWAPS_PERFORMANCE_SCENARIO_001.slug
+  const runId = `${scenario.metadata.id.toLowerCase()}-${
+    scenario.metadata.slug
   }-${createdAt.toISOString().replace(/[:.]/gu, '-')}`;
-  const phases: ScenarioPhase[] = [];
+  let phases: ScenarioPhase[] = [];
+  let preconditions: ScenarioPreconditionState = { walletUnlocked: false };
   let capture: RuntimeCapture | null = null;
-  let sourceTokenText: string | null = null;
-  let walletUnlocked = false;
   let failure: string | null = null;
   let diagnosticsInstalled = false;
   let sessionAvailable = false;
@@ -235,6 +193,13 @@ async function runScenario(): Promise<void> {
   log(`using Metro port ${metroPort}`);
 
   try {
+    const instrumentationStatus = getInstrumentationStatus(process.cwd());
+    if (instrumentationStatus !== 'prepared') {
+      throw new Error(
+        `Instrumentation profile ${scenario.metadata.instrumentationProfile} is ${instrumentationStatus}; run yarn performance:swaps prepare before Metro starts`,
+      );
+    }
+
     log('checking iOS simulator prerequisites');
     runYarn(['mm:doctor']);
 
@@ -244,7 +209,7 @@ async function runScenario(): Promise<void> {
       throw new Error(
         `No active mm session is available. First run ${formatMmSessionSetupCommand(
           metroPort,
-        )}, then unlock MetaMask and leave it on Wallet before rerunning ${SWAPS_PERFORMANCE_SCENARIO_001.id}.`,
+        )}, then unlock MetaMask and leave it on Wallet before rerunning ${scenario.metadata.id}.`,
       );
     }
     sessionAvailable = true;
@@ -252,8 +217,14 @@ async function runScenario(): Promise<void> {
       'reusing the active mm session without launching or refreshing the app',
     );
 
-    waitForTestId(WALLET_SWAP_BUTTON_TEST_ID, 10_000);
-    walletUnlocked = true;
+    runMm([
+      'wait-for',
+      '--testid',
+      'homepage-action-buttons-grid-swap',
+      '--timeout',
+      '10000',
+    ]);
+    preconditions = { walletUnlocked: true };
 
     const installResult = evaluateRuntime(buildInstallDiagnosticsExpression());
     if (typeof installResult !== 'string') {
@@ -261,46 +232,14 @@ async function runScenario(): Promise<void> {
     }
     diagnosticsInstalled = true;
 
-    log('opening Swaps');
-    phases.push(
-      await measurePhase('open-swaps', () => {
-        clickTestId(WALLET_SWAP_BUTTON_TEST_ID);
-        waitForTestId('source-token-area-input', 15_000);
-      }),
-    );
-
-    sourceTokenText = getExactScreenText('source-token-selector-button');
-    if (!sourceTokenText?.toUpperCase().includes('ETH')) {
-      throw new Error(
-        'Expected ETH as the source token. Switch the Wallet to Ethereum and retry.',
-      );
-    }
-
-    log('selecting Ethereum USDC as the destination');
-    phases.push(
-      await measurePhase('select-destination', () => {
-        clickTestId('dest-token-selector-button');
-        // USDC is a prioritized Ethereum asset and has a unique chain-scoped ID.
-        // Selecting it directly avoids the generic iOS TextField wrapper ID.
-        waitForTestId('asset-0x1-USDC', 20_000);
-        clickTestId('asset-0x1-USDC');
-        waitForTestId('dest-token-area-input', 10_000);
-      }),
-    );
-
-    log('entering 1 ETH and waiting for the first quote');
-    phases.push(
-      await measurePhase('fetch-first-quote', async () => {
-        clickTestId('source-token-area-input');
-        waitForTestId('keypad-delete-button', 10_000);
-        clickTestId('keypad-delete-button');
-        clickTestId('keypad-key-1');
-        await waitForPositiveQuote();
-      }),
-    );
+    const result = await scenario.run(createScenarioContext());
+    phases = result.phases;
+    preconditions = result.preconditions;
 
     markRuntime('scenario:complete');
-    await delay(500);
+    await new Promise((resolveDelay) => {
+      setTimeout(resolveDelay, 500);
+    });
     runMm(['describe-screen']);
     runMm(['screenshot', '--name', `${runId}-quote`]);
     log('quote became visible');
@@ -319,32 +258,27 @@ async function runScenario(): Promise<void> {
 
     if (!failure && (!capture || Object.keys(capture.renders).length === 0)) {
       failure =
-        'No render probes were captured. Run yarn performance:swaps:prepare before starting the scenario.';
+        'No render probes were captured. Run yarn performance:swaps prepare before starting the scenario.';
     }
 
-    const summary = capture ? summarizeCapture(capture, phases) : null;
     const artifact: SwapsPerformanceArtifact = {
       schemaVersion: 1,
       run: {
         id: runId,
-        scenario: SWAPS_PERFORMANCE_SCENARIO_001.slug,
-        scenarioId: SWAPS_PERFORMANCE_SCENARIO_001.id,
-        scenarioName: SWAPS_PERFORMANCE_SCENARIO_001.name,
+        scenario: scenario.metadata.slug,
+        scenarioId: scenario.metadata.id,
+        scenarioName: scenario.metadata.name,
+        scenarioDescription: scenario.metadata.description,
         createdAt: createdAt.toISOString(),
         commit: getCommit(),
-        platform: 'ios-simulator',
+        platform: scenario.metadata.platform,
         metroPort,
         status: failure ? 'failed' : 'passed',
       },
-      preconditions: {
-        walletUnlocked,
-        sourceTokenText,
-        destinationToken: 'USDC',
-        sourceAmount: '1',
-      },
+      preconditions,
       phases,
       capture,
-      summary,
+      summary: capture ? summarizeCapture(capture, phases) : null,
       failure,
     };
     const paths = writeArtifact(artifact);
@@ -360,9 +294,3 @@ async function runScenario(): Promise<void> {
     throw new Error(failure);
   }
 }
-
-runScenario().catch((error: unknown) => {
-  const message = error instanceof Error ? error.message : String(error);
-  process.stderr.write(`${LOG_PREFIX} ${sanitizeFailure(message)}\n`);
-  process.exitCode = 1;
-});
